services/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standart"], version = "^0.28.0"}
pydantic-settings = "^2.2.1"
pydantic = "^2.6.3"
sqlalchemy = "^2.0.28"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
